Global exception added

+++ src/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Product, ProductDocument } from '../schemas/product/product.schema';
@@ -57,11 +57,24 @@ export class ProductsService {
   // Get single product
   async findOne(param: ProductIDDto) {
     const { id } = param;
-    return this.productModel.findById(id).exec();
+
+    const product = await this.productModel.findById(id).exec();
+
+    if (!product) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
+    }
+
+    return product;
   }
 
   // Update product
   async update(id: string, product: CreateProductDto) {
+    const findProduct = await this.productModel.findById(id).exec();
+
+    if (!findProduct) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
+    }
+
     return this.productModel
       .findByIdAndUpdate(id, product, { new: true })
       .exec();
@@ -70,6 +83,12 @@ export class ProductsService {
   // Delete Product
   async remove(params: ProductIDDto) {
     const { id } = params;
+    const findProduct = await this.productModel.findById(id).exec();
+
+    if (!findProduct) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
+    }
+
     return this.productModel.findByIdAndDelete(id).exec();
   }
 }
